+++ elfo-core/src/time/interval.rs
@@ -12,7 +12,7 @@ use crate::{
     addr::Addr,
     envelope::{Envelope, MessageKind},
     message::Message,
-    trace_id,
+    tracing::TraceId,
 };
 
 pub struct Interval<F> {
@@ -99,7 +99,7 @@ where
             // Emit a message.
             let message = (self.message_factory)();
             let kind = MessageKind::Regular { sender: Addr::NULL };
-            let trace_id = trace_id::generate();
+            let trace_id = TraceId::generate();
             let envelope = Envelope::with_trace_id(message, kind, trace_id).upcast();
             return Poll::Ready(Some(envelope));
         }